rds: add rds management
