+++ package.json
@@ -6,8 +6,8 @@
   "bugs": "https://github.com/pagarme/mr-krabs/issues",
   "main": "lib/index.js",
   "repository": {
-    "type" : "git",
-    "url" : "https://github.com/pagarme/mr-krabs.git"
+    "type": "git",
+    "url": "https://github.com/pagarme/mr-krabs.git"
   },
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1"
